--- honeycomb-core/src/cells/mod.rs
@@ -1,8 +1,0 @@
-//! Cell modeling code
-//!
-//! This module contains all code related to cell & orbit modeling.
-
-// ------ MODULE DECLARATIONS
-
-pub mod collections;
-pub mod orbits;
