+++ honeycomb-core/src/cmap/builder/grid/descriptor.rs
