+++ honeycomb-core/src/cmap/dim2/basic_ops.rs
