+++ honeycomb-core/src/cmap/dim2/embed.rs
