+++ honeycomb-core/src/spatial_repr/mod.rs
@@ -5,10 +5,23 @@
 
 // ------ MODULE DECLARATIONS
 
-pub mod vector;
-pub mod vertex;
+mod vector;
+mod vertex;
+
+// ------ IMPORTS
+
+use std::any::TypeId;
+use std::fmt::Debug;
+use std::ops::{AddAssign, DivAssign, MulAssign, SubAssign};
 
 // ------ CONTENT
+
+// --- re-exports
+
+pub use vector::Vector2;
+pub use vertex::Vertex2;
+
+// --- error enum
 
 /// Coordinates-level error enum
 #[derive(Debug, PartialEq)]
@@ -17,6 +30,24 @@ pub enum CoordsError {
     ///
     /// This is returned when trying to compute the unit vector of a null [`vector::Vector2`].
     InvalidUnitDir,
+}
+
+// --- generic fp repersentation trait
+
+/// Common trait implemented by types used for coordinate representation.
+///
+/// The static lifetime is a requirements induced by specific implementations that use [`TypeId`];
+/// This is used in order to identify types in two contexts:
+/// - Interacting with VTK files (`io` feature),
+/// - Coding vertices and generic attributes handling
+pub trait CoordsFloat:
+    num::Float + Default + AddAssign + SubAssign + MulAssign + DivAssign + Debug + 'static
+{
+}
+
+impl<T: num::Float + Default + AddAssign + SubAssign + MulAssign + DivAssign + Debug + 'static>
+    CoordsFloat for T
+{
 }
 
 // ------ TESTS
